+++ liquibase/observe/sql/tables/fact_file_processings_00.sql
@@ -2,11 +2,11 @@ CREATE TABLE fact_file_processings
 (
     id                 BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
     dim_version_id     INT UNSIGNED    NOT NULL,
-    dim_file_id        BIGINT UNSIGNED NOT NULL,
+    dim_file_id        int UNSIGNED    NOT NULL,
     dim_stage_id       INT UNSIGNED    NOT NULL,
     dim_status_id      INT UNSIGNED    NOT NULL,
     process_timestamp  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    dim_batch_id       BIGINT UNSIGNED NULL,
+    dim_batch_id       int UNSIGNED    NULL,
     error_message      TEXT            NULL,
     processing_time_ms INT UNSIGNED    NULL,
     created_at         TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
